+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/ui/ShowsActivity.java
@@ -62,6 +62,8 @@ import com.battlelancer.seriesguide.sync.AccountUtils;
 import com.battlelancer.seriesguide.sync.SgSyncAdapter;
 import com.battlelancer.seriesguide.ui.FirstRunFragment.OnFirstRunDismissedListener;
 import com.battlelancer.seriesguide.ui.dialogs.AddDialogFragment;
+import com.battlelancer.seriesguide.ui.streams.FriendsEpisodeStreamFragment;
+import com.battlelancer.seriesguide.ui.streams.UserEpisodeStreamFragment;
 import com.battlelancer.seriesguide.util.DBUtils;
 import com.battlelancer.seriesguide.util.EpisodeTools;
 import com.battlelancer.seriesguide.util.ImageProvider;
@@ -84,7 +86,12 @@ public class ShowsActivity extends BaseTopShowsActivity implements
 
     protected static final String TAG = "Shows";
 
-    public static final int ADD_SHOW_LOADER_ID = 1;
+    public static final int SHOWS_LOADER_ID = 100;
+    public static final int UPCOMING_LOADER_ID = 101;
+    public static final int RECENT_LOADER_ID = 102;
+    public static final int FRIENDS_LOADER_ID = 103;
+    public static final int USER_LOADER_ID = 104;
+    public static final int ADD_SHOW_LOADER_ID = 105;
 
     private static final int UPDATE_SUCCESS = 100;
 
@@ -227,7 +234,7 @@ public class ShowsActivity extends BaseTopShowsActivity implements
         argsUpcoming.putString(ActivityFragment.InitBundle.TYPE,
                 ActivityFragment.ActivityType.UPCOMING);
         argsUpcoming.putString(ActivityFragment.InitBundle.ANALYTICS_TAG, "Upcoming");
-        argsUpcoming.putInt(ActivityFragment.InitBundle.LOADER_ID, 10);
+        argsUpcoming.putInt(ActivityFragment.InitBundle.LOADER_ID, UPCOMING_LOADER_ID);
         argsUpcoming.putInt(ActivityFragment.InitBundle.EMPTY_STRING_ID, R.string.noupcoming);
         mTabsAdapter.addTab(R.string.upcoming, ActivityFragment.class, argsUpcoming);
 
@@ -236,13 +243,14 @@ public class ShowsActivity extends BaseTopShowsActivity implements
         argsRecent
                 .putString(ActivityFragment.InitBundle.TYPE, ActivityFragment.ActivityType.RECENT);
         argsRecent.putString(ActivityFragment.InitBundle.ANALYTICS_TAG, "Recent");
-        argsRecent.putInt(ActivityFragment.InitBundle.LOADER_ID, 20);
+        argsRecent.putInt(ActivityFragment.InitBundle.LOADER_ID, RECENT_LOADER_ID);
         argsRecent.putInt(ActivityFragment.InitBundle.EMPTY_STRING_ID, R.string.norecent);
         mTabsAdapter.addTab(R.string.recent, ActivityFragment.class, argsRecent);
 
-        // trakt friends tab
+        // trakt tabs only visible if connected
         if (TraktCredentials.get(this).hasCredentials()) {
-            mTabsAdapter.addTab(R.string.friends, TraktFriendsFragment.class, null);
+            mTabsAdapter.addTab(R.string.friends, FriendsEpisodeStreamFragment.class, null);
+            mTabsAdapter.addTab(R.string.user_stream, UserEpisodeStreamFragment.class, null);
         }
 
         // display new tabs
@@ -431,26 +439,9 @@ public class ShowsActivity extends BaseTopShowsActivity implements
     }
 
     @Override
-    public boolean onPrepareOptionsMenu(Menu menu) {
-        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
-
-        // If the nav drawer is open, hide action items related to the content
-        // view
-        boolean isDrawerOpen = isDrawerOpen();
-        menu.findItem(R.id.menu_add_show).setVisible(!isDrawerOpen);
-
-        return super.onPrepareOptionsMenu(menu);
-    }
-
-    @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         int itemId = item.getItemId();
-        if (itemId == R.id.menu_add_show) {
-            fireTrackerEvent("Add show");
-            startActivity(new Intent(this, AddActivity.class));
-            overridePendingTransition(R.anim.fade_in, R.anim.fade_out);
-            return true;
-        } else if (itemId == R.id.menu_search) {
+        if (itemId == R.id.menu_search) {
             startActivity(new Intent(this, SearchActivity.class));
             fireTrackerEvent("Search");
             return true;
@@ -594,7 +585,7 @@ public class ShowsActivity extends BaseTopShowsActivity implements
     }
 
     /**
-     * Called if the user adds a show from {@link com.battlelancer.seriesguide.ui.TraktFriendsFragment}.
+     * Called if the user adds a show from {@link FriendsEpisodeStreamFragment}.
      */
     @Override
     public void onAddShow(SearchResult show) {